refactor: added header file, split lexer and parser
